Testing azuredeployment

+++ CustomerAppBLL/Services/CustomerService.cs
@@ -56,7 +56,8 @@ namespace RestAppBLL.Services
         {
             using (var uow = _dalFacade.UnitOfWork)
             {
-                return _converter.Convert(uow.CustomerRepository.GetById(id));
+                var customer = uow.CustomerRepository.GetById(id);
+                return customer == null ? null : _converter.Convert(customer);
             }
         }
 
@@ -64,6 +65,8 @@ namespace RestAppBLL.Services
         {
             using (var uow = _dalFacade.UnitOfWork)
             {
+                var customerFromDB = uow.CustomerRepository.GetById(id);
+                if (customerFromDB == null) return false;
                 var customerDeleted = uow.CustomerRepository.Delete(id);
                 uow.Complete();
                 return customerDeleted;
@@ -75,7 +78,7 @@ namespace RestAppBLL.Services
             using (var uow = _dalFacade.UnitOfWork)
             {
                 var customerFromDb = uow.CustomerRepository.GetById(updatedCustomer.Id);
-                if (customerFromDb == null) throw new InvalidOperationException("Customer doesn't exist in DB");
+                if (customerFromDb == null) return null;
 
                 customerFromDb.FirstName = updatedCustomer.FirstName;
                 customerFromDb.LastName = updatedCustomer.LastName;
